Refactor address manipulation functions in a separate d.comm.addressing module
