+++ distributed/cli/utils.py
@@ -12,8 +12,8 @@ For more information see: http://click.pocoo.org/5/python3/
 """.strip()
 
 
-from distributed.comm.core import (parse_address, unparse_address,
-                                   parse_host_port, unparse_host_port)
+from distributed.comm import (parse_address, unparse_address,
+                              parse_host_port, unparse_host_port)
 from ..utils import get_ip, ensure_ip
